awd_django/requirements.txt
Django==4.2.19
djangorestframework==3.15.2
django-cors-headers==4.4.0
djoser==2.3.1
django-apscheduler==0.7.0
psycopg2==2.9.10
celery==5.4.0
redis==5.2.1
django-redis==5.4.0
whitenoise==6.6.0
gunicorn==20.1.0
djangorestframework-simplejwt==5.3.1